Update the claroc.sh CLI Script to Accept Paths to Necessary Jars Via Env Vars
This change makes it so that the script doesn't just work on my machine anymore. Now, you should be able to "simply" configure the paths to `claro_compiler_binary_deploy.jar` and `claro_builtin_java_deps_deploy.jar` as ENV vars `CLARO_COMPILER_PATH` and `CLARO_BUILTIN_DEPS_PATH` respectfully. For now, you get the first Jar by simply grabbing it from the latest Github release on clarolang.com, but the deps Jar still needs to be setup for releases in the Github release action. That's still a TODO.

The nice thing here is that until I actually manage to find a more legit approach to "installing" these CLIs, once I get these CLIs and the necessary Jars included as part of the Github Actions release process, you should be able to simply download the release zip, unzip and then start compiling from that directory w/o even specifying any env vars as the script will auto-check for the jars in the curr dir.

This current script should be ready for use w/in the Riju configuration to get Riju updates back to a working state.

+++ cli/claroc.sh
@@ -18,40 +18,57 @@ function main() {
   # Move into a tempdir so that everything can be cleaned up in the end. The only artifact I want this script to leave
   # behind is literally just the fat executable Jar containing the compiled Claro program.
   ORIG_DIR=$(pwd)
-  tempfoo=$(basename $0)
+  tempfoo=$(basename "$0")
   TMP_DIR=$(mktemp -d -t "${tempfoo}") || exit 1
   echo "Created tempdir: $TMP_DIR"
   cp "$@" "$TMP_DIR"
   cd "$TMP_DIR" || die
 
-  echo "Creating symlink to Claro Compiler Jar"
   CLARO_COMPILER_JAR_NAME="claro_compiler_binary_deploy.jar"
-  # TODO(steving) I need to figure out how to consistently identify this file's location on various machines...
-  CLARO_COMPILER_PATH="/private/var/tmp/_bazel_jasonsteving/efcd1bf992362b57bda2d1a8112007a7/execroot/claro-lang/bazel-out/host/bin/src/java/com/claro/claro_compiler_binary_deploy.jar"
-  maybeCreateSymlink $CLARO_COMPILER_JAR_NAME $CLARO_COMPILER_PATH
+  # TODO(steving) I need to figure out better way to consistently identify this file's location on various machines...
+  # Check if CLARO_COMPILER_PATH is defined
+  if CLARO_COMPILER_PATH=$(check_env_var_and_set_default "\$CLARO_COMPILER_PATH" "$CLARO_COMPILER_PATH" $CLARO_COMPILER_JAR_NAME);
+  then
+    # Continue with the rest of the script
+    echo "\$CLARO_COMPILER_PATH is set to $CLARO_COMPILER_PATH"
+  else
+    echo "$CLARO_COMPILER_PATH" >&2 && die
+  fi
 
   echo "Creating symlink to Claro Builtin Deps Jar"
   CLARO_BUILTIN_DEPS_JAR_NAME="claro_builtin_java_deps_deploy.jar"
-  # TODO(steving) I need to figure out how to consistently identify this file's location on various machines...
-  CLARO_BUILTIN_DEPS_PATH="/private/var/tmp/_bazel_jasonsteving/efcd1bf992362b57bda2d1a8112007a7/execroot/claro-lang/bazel-out/darwin-fastbuild/bin/src/java/com/claro/claro_programs/claro_builtin_java_deps_deploy.jar"
-  maybeCreateSymlink $CLARO_BUILTIN_DEPS_JAR_NAME $CLARO_BUILTIN_DEPS_PATH
+  # TODO(steving) I need to figure out better way to consistently identify this file's location on various machines...
+  # Check if CLARO_BUILTIN_DEPS_PATH is defined
+  if CLARO_BUILTIN_DEPS_PATH=$(check_env_var_and_set_default "\$CLARO_BUILTIN_DEPS_PATH" "$CLARO_BUILTIN_DEPS_PATH" $CLARO_BUILTIN_DEPS_JAR_NAME);
+  then
+    # Continue with the rest of the script
+    echo "\$CLARO_BUILTIN_DEPS_PATH is set to $CLARO_BUILTIN_DEPS_PATH"
+  else
+    echo "Error: $CLARO_BUILTIN_DEPS_PATH" >&2  && die
+  fi
 
-  # TODO(steving) I need to figure out how to consistently identify this file's location on various machines...
-  CLARO_STDLIB_SRCS="/Users/jasonsteving/Claro/claro-lang/src/java/com/claro/stdlib/claro/builtin_functions.claro_internal,/Users/jasonsteving/Claro/claro-lang/src/java/com/claro/stdlib/claro/builtin_types.claro_internal"
+  # I've intentionally packed the stdlib .claro_internal srcs into the claro_builtin_java_deps_deploy.jar so I can
+  # extract them from the Jar to work with here.
+  jar -xf "$CLARO_BUILTIN_DEPS_PATH" com/claro/stdlib/claro/
+  # Dynamically lookup the .claro_internal files found packed in the jar so I don't need to keep updating this script.
+  CLARO_STDLIB_SRCS=$(find com -type f -name "*.claro_internal")
+  CLARO_STDLIB_SRCS=$(echo "$CLARO_STDLIB_SRCS" | tr '\n' ',') # Has trailing comma...
 
   echo "Compiling $CLARO_MAIN_FILE.claro..."
-  java -jar claro_compiler_binary_deploy.jar --java_source --silent=true --classname=$CLARO_MAIN_FILE --package=com.claro --srcs=$CLARO_STDLIB_SRCS,$CLARO_FILE_NAMES > $CLARO_MAIN_FILE.java || die
+  java -jar "$CLARO_COMPILER_PATH" --java_source --silent=true --classname="$CLARO_MAIN_FILE" --package=com.claro --srcs="$CLARO_STDLIB_SRCS""$CLARO_FILE_NAMES" > "$CLARO_MAIN_FILE".java || die
 
   echo "Compiling $CLARO_MAIN_FILE.java..."
-  javac -cp ".:claro_builtin_java_deps_deploy.jar" "$CLARO_MAIN_FILE".java -d .
+  javac -cp ".:$CLARO_BUILTIN_DEPS_PATH" "$CLARO_MAIN_FILE".java -d .
 
   # Here, I'm creating a "fat jar" executable so that there's a single executable artifact produced by this script
   # which you should be able to take anywhere that Java is installed and run it on its own.
   echo "Generating Executable $CLARO_MAIN_FILE.jar..."
-  cp $CLARO_BUILTIN_DEPS_JAR_NAME "$CLARO_MAIN_FILE".jar
+  cp "$CLARO_BUILTIN_DEPS_PATH" "$CLARO_MAIN_FILE".jar
   jar -uf "$CLARO_MAIN_FILE".jar com/claro/*.class
   # I need to overwrite the entrypoint of the jarfile so that it starts at the codegen'd java for the user's Claro program.
   jar -ufe "$CLARO_MAIN_FILE".jar com.claro."$CLARO_MAIN_FILE"
+  # This executable Jar doesn't need the stdlib sources anymore. Remove them. Remember that Jars are just zip files.
+  zip -d "$CLARO_MAIN_FILE".jar "com/claro/stdlib/claro/*" > /dev/null || true
 
   echo "Moving executable Jar out of tmpdir."
   mv "$CLARO_MAIN_FILE".jar "$ORIG_DIR"
@@ -60,14 +77,14 @@ function main() {
   cd "$ORIG_DIR" || die
   rm -r "$TMP_DIR"
 
-  echo "Done. Run your program using: \"claroc.sh $CLARO_MAIN_FILE\""
+  echo "Done. Run your program using: \"claro.sh $CLARO_MAIN_FILE\""
 }
 
 function validateArgs() {
   # Make sure there were actually some args passed.
   if [ $# -eq 0 ]; then
-    echo "claroc requires at least one .claro file to compile. Try rerunning like the following:"
-    echo "  $ ./claroc.sh foo.claro [...]"
+    echo "claroc requires at least one .claro file to compile. Try rerunning like the following:" >&2
+    echo "  $ ./claroc.sh foo.claro [...]" >&2
     exit 1
   fi
 
@@ -85,23 +102,36 @@ function validateArgs() {
   done
 }
 
-function maybeCreateSymlink() {
-  symlink_name="$1"
-  symlink_full_path="$2"
-  # Check if the symlink already exists
-  if [ ! -L "$symlink_name" ]; then
-    # The symlink doesn't exist, so create it
-    ln -s "$symlink_full_path" .
-  else
-    # The symlink already exists, so print a message
-    echo "Symlink to $symlink_name already exists."
+function check_env_var_and_set_default() {
+  # Get the name of the environment variable
+  local env_var_val=$2
+  # Get the name of the file to look for
+  local file_name=$3
+
+  # Check if the environment variable is defined
+  if [ -z "$env_var_val" ]; then
+    # Check if the file exists in the current directory
+    if [ -f "$ORIG_DIR/$file_name" ]; then
+      # Set the environment variable to the path of the file
+      env_var_val="$ORIG_DIR/$file_name"
+    else
+      # Exit because the environment variable is not defined and the file does not exist
+      echo "The environment variable $1 is not defined and the file $file_name does not exist. Exiting." >&2
+      exit 1
+    fi
   fi
+  # Return the value of the environment variable
+  echo "$env_var_val"
+  exit 0
 }
+
 
 function die() {
   # Make sure to still cleanup the tmpdir
-  rm -r "$TMP_DIR"
+  if [[ -z "$TMP_DIR" ]]; then
+    rm -r "$TMP_DIR"
+  fi
   exit 1
 }
 
-main "$@"; exit
+main "$@"; die
